// angular-components-oasis/projects/oasis-short-calendar/src/shared/short-calendar-date.interface.ts
/**
 * Интерфейс для хранения даты
 */
export interface ShortCalendarDate {
  /** Первый месяц (единичная или первая дата) */
  month: number;

  /** Номер дня (единичная или первая дата) */
  dayNumber: number;
}